+++ collective/honeypot/interfaces.py
@@ -3,8 +3,13 @@ from zope.interface import Interface
 
 
 class IHoneypot(Interface):
-    """Honeypot text field.
-    """
+    """Honeypot text field."""
 
     # Keep field title empty so visitors do not see it.
     honeypot = schema.TextLine(title=u"", required=False)
+
+
+class IHoneypotDisabledForm(Interface):
+    """
+    Marker interface to disable authomatic field insertion in some views
+    """